rename `utils/windows` into `core/windows`
