--- utils/windows/CMakeLists.txt
@@ -1,2 +1,0 @@
-add_library(volume_utils_windows INTERFACE)
-target_include_directories(volume_utils_windows INTERFACE include)
